refactor(app): reorganize model imports and update routes for improved structure
