+++ app/models/entities/PartialRoom.py
