+++ app/templates/admin/manage-rooms.html
