+++ app/templates/mails/register.html
