+++ app/templates/rooms/room.html
